fix(ui): tickets: id column width
Problem:
Id column width in tickets view is set as a constant value. (2
currently). While this works for out of the box setups, it might not
work for dynamic formats (eg: expression).

Investigation:
Interestingly, this used to be dynamic
```
width: manager.totalCount > 999 ? '2' : '1'
```
This behavior was changed in 7329f863bdfb553c283627078faa3269ea130ed6.
While this is not perfect, it could have worked for at-least some
cases.

Solution:
I introduced another logic to calculate width, depending on the widest
(longest string) id from the table. It is safe to assume that all ids
will be of same length, and this logic will work for most cases. There
could be room for improvement though, as more cases pop up.